+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 =====
 
 * (improvement) Replace `.` with `-` in `TaskMetaData::getKey`.
+* (improvement) Change column of `TaskRun::$taskLog` to be not nullable.
 * (internal) Update `21torr/janus` and use `phpunit` instead of `simple-phpunit`.